Contacts: add basic sync support with vdirsyncer, split AddressBook and Contact
